==> notebooks/01_MainNB.ipynb <==
# Function to Automate Web Scraping of Historical Yellow Cab Rides in NYC

import requests
from pathlib import Path

def DownloadOneFileRawData(year:int, month:int) -> Path:
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(URL)

    if response.status_code == 200:
        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        open(path, "wb").write(response.content)
        return path
    else:
        raise Exception(f"{URL} is not available")

DownloadOneFileRawData(2022, 1)

import pandas as pd

rides = pd.read_parquet("../data/raw/rides_2022-01.parquet")

rides.head(10)

rides = rides[["tpep_pickup_datetime", "PULocationID"]]

rides.rename(columns={"tpep_pickup_datetime":"pickup_datetime", "PULocationID":"pickup_location"}, inplace = True)

rides

# Data Validation - Checking that all Pickups are actually from the Month they are suppose to be

rides["pickup_datetime"].describe()

rides = rides[rides["pickup_datetime"] >= "2022-01-01"]
rides = rides[rides["pickup_datetime"] <= "2022-02-01"]

rides["pickup_datetime"].describe()

#Writing down on Disk the Transformed and Validated Data

rides.to_parquet("../data/transformed/rides_2022-01.parquet")

# Data Transformation - Transforming Raw Data into Time-Series Data

import pandas as pd

newrides = pd.read_parquet("../data/transformed/rides_2022-01.parquet")

newrides.head(10)

newrides["pickup_hour"] = newrides["pickup_datetime"].dt.floor('H')

newrides

aggrides = newrides.groupby(["pickup_hour", "pickup_location"]).size().reset_index()
aggrides.rename(columns = {0 : "numrides"}, inplace = True)
aggrides

#Checking for 0 Rides Rows and Adding Empty Hours when No Rides Occurred

from tqdm import tqdm

def AddMissingSlots(aggrides:pd.DataFrame) -> pd.DataFrame:
    
    locations = aggrides["pickup_location"].unique()
    full_range = pd.date_range(aggrides["pickup_hour"].min(), aggrides["pickup_hour"].max(), freq = "H")
    output = pd.DataFrame()
    
    for locid in tqdm(locations):
    
        #Keep only Rides for this Location ID
        aggrides_i = aggrides.loc[aggrides["pickup_location"] == locid, ["pickup_hour", "numrides"]]
        
        #Adding Missing Dates with 0 in a Series
        aggrides_i.set_index("pickup_hour", inplace = True)
        aggrides_i.index = pd.DatetimeIndex(aggrides_i.index)
        aggrides_i = aggrides_i.reindex(full_range, fill_value = 0)
        
        #Add Back Location ID Columns
        aggrides_i["pickup_location_id"] = locid
        
        output = pd.concat([output, aggrides_i])
        
    #Move the PickupHour from Index to Column
    output = output.reset_index().rename(columns = {"index":"pickup_hour"})
    
    return output

#Creating New Complete Time-Series DF

AggRides = AddMissingSlots(aggrides)

AggRides

# Plotting Data

from typing import Optional, List
import plotly.express as px

def PlotRides(rides:pd.DataFrame, locations:Optional[List[int]] = None):
    
    rides_to_plot = rides[rides["pickup_location_id"].isin(locations)] if locations else rides
    
    fig = px.line(
        rides_to_plot,
        x = "pickup_hour",
        y = "numrides",
        color = "pickup_location_id",
        template = "none",
    )
    
    fig.show()

# The Following Import and Setting is needed to Make the Plotly Figure Showing
import plotly.io as pio
pio.renderers.default = "iframe" # or 'colab' or 'iframe' or 'iframe_connected' or 'sphinx_gallery'

PlotRides(AggRides, locations = [43])

#Saving the Data to Disk

AggRides.to_parquet("../data/transformed/TransformedData_2022-01.parquet")

# Transforming the Time-Series in Training Data (Tabular Data)

#The Features are going to be an X Number of Rows, each one counting as a Feature
#The Target is going to be the Next Immediate Row after the Ones selected as Features
#On Every new Target to Predict we Slide the Rows by 1
#So, the Features on Each iteration are always going to be an X Number of Rows.
#(The Previous Target becomes a Feature when Predicting the Next Target and we Drop the First Row Used)

#EXAMPLE:
#Rolling 12 Features, we Take the first 12 Rows to predict the 13th as a Target
#Features Rows Indexes = 0 - 12
#Target Row Index = 13
#NEXT ITERATION
#We Take the 12 Rows from Index 1 to 13 to predict the 14th as a Target
#Features Rows Indexes = 1 - 13
#Target Row Index = 14

TransformedData = pd.read_parquet("../data/transformed/TransformedData_2022-01.parquet")
TransformedData

#Just Picking a One Location now

TS_OneLocation = TransformedData.loc[TransformedData["pickup_location_id"] == 43, :].reset_index(drop = True)
TS_OneLocation

def GetCutoffIndeces(data:pd.DataFrame, nFeatures:int, SlidingFactor:int) -> list:
    StopPosition = len(data)-1
    
    #Start the First SubSequence at Index Position 0
    SubseqFirstIdx = 0
    SubseqStepIdx = nFeatures
    SubseqLastIdx = nFeatures +1
    
    #[FirstIdx, StepIdx, LastIdx]
    
    Indeces = []
    
    while SubseqLastIdx <= StopPosition:
        Indeces.append([SubseqFirstIdx, SubseqStepIdx, SubseqLastIdx])
        
        #StepSize is used as Sliding Factor
        SubseqFirstIdx += SlidingFactor
        SubseqStepIdx += SlidingFactor
        SubseqLastIdx += SlidingFactor
        
    return Indeces

nFeatures = 24
SlidingFactor = 1

Indeces = GetCutoffIndeces(TS_OneLocation, nFeatures, SlidingFactor)

Indeces[:5]

#Implementing the Slicing using the Generated Indeces

import numpy as np
import warnings
warnings.filterwarnings("ignore") 

nSamples = len(Indeces)
X = np.ndarray(shape = (nSamples, nFeatures), dtype = np.float32)
Y = np.ndarray(shape = (nSamples), dtype = np.float32)

PickupHours = []

for i, idx in enumerate(Indeces):
    X[i,:] = TS_OneLocation.iloc[idx[0]:idx[1]]["numrides"].values
    Y[i] = TS_OneLocation.iloc[idx[1]:idx[2]]["numrides"].values
    PickupHours.append(TS_OneLocation.iloc[idx[1]]["pickup_hour"])

#Constructing a Features DataFrame with the Columns name inteded as Number of Rides X Hours Before the Targets' Hour

FeaturesOneLocationDF = pd.DataFrame(X, columns = [f"rides_{i+1}_hours_before" for i in reversed(range(nFeatures))])
FeaturesOneLocationDF

#Constructing a Targets DataFrame with the Columns name inteded as Number of Rides AT the Targets' Hour

TargetsOneLocationDF = pd.DataFrame(Y, columns = ["target_rides_next_hour"])
TargetsOneLocationDF

#Defining a Function that Does the Trick for all Locations and Not Just One
#(Transforms all Time-Series for Each of ALL Locations into Tabular Format as Above)

def TransformALL(tsData:pd.DataFrame, nFeatures:int, SlidingFactor:int) -> pd.DataFrame:
    
    assert set(tsData.columns) == {"pickup_hour", "numrides", "pickup_location_id"}
    
    locationIDs = tsData["pickup_location_id"].unique()
    Features = pd.DataFrame()
    Targets = pd.DataFrame()
    
    for locid in tqdm(locationIDs):
        #Keep only Time-Series Data for this Location
        tsDataOneLocation = tsData.loc[tsData["pickup_location_id"] == locid, ["pickup_hour", "numrides"]].sort_values(by = ["pickup_hour"])
        
        #Pre-Compute Cutoff Indeces to Split DataFrame Rows
        indeces = GetCutoffIndeces(tsDataOneLocation, nFeatures, SlidingFactor)
        
        #Slice and Transpose Data into NumPy Arrays for Features and Target
        nSamples = len(indeces)
        X = np.ndarray(shape=(nSamples, nFeatures), dtype=np.float32)
        Y = np.ndarray(shape=(nSamples), dtype=np.float32)
        PickupHours = []
        
        for i, idx in enumerate(indeces):
            X[i,:] = tsDataOneLocation.iloc[idx[0]:idx[1]]["numrides"].values
            Y[i] = tsDataOneLocation.iloc[idx[1]:idx[2]]["numrides"].values
            PickupHours.append(tsDataOneLocation.iloc[idx[1]]["pickup_hour"])
            
        #NumPy -> Pandas
        FeaturesOneLocationDF = pd.DataFrame(X, columns = [f"rides_{i+1}_hours_before" for i in reversed(range(nFeatures))])
        FeaturesOneLocationDF["pickup_hour"] = PickupHours
        FeaturesOneLocationDF["pickup_location_id"] = locid
        
        TargetsOneLocationDF = pd.DataFrame(Y, columns = ["target_rides_next_Hour"])

        #Concatenate Results
        Features = pd.concat([Features, FeaturesOneLocationDF])
        Targets = pd.concat([Targets, TargetsOneLocationDF])
        
    Features.reset_index(inplace = True, drop = True)
    Targets.reset_index(inplace = True, drop = True)
    
    return Features, Targets["target_rides_next_Hour"]

Features, Targets = TransformALL(TransformedData, nFeatures = 24*7*1, SlidingFactor = 24) #One Week of Features

Features
