==> notebooks/04_BackFillFeaturesStore.ipynb <==
HopsworksProjectName = "TaxiDemandPrediction"

import os
import sys
sys.path.append("../src/")
sys.path.append("../")
import config

import APIKey

HopsworksAPIKey = APIKey.HOPSWORKSAPIKEY

# Fetching Historical Data from 2022 Onwards from the TLC Trip Data 

from datetime import datetime 
import pandas as pd
sys.path.append("../src/")
import data

StartYear = 2022
EndYear = datetime.now().year
print(f"Downloading Data from Year {StartYear} to Year {EndYear}")

Rides = pd.DataFrame()

for y in range(StartYear, EndYear+1):
    
    #Download one whole Year worth of Data
    RidesOneWholeYear = data.LoadRawData(y)
    
    #Append Rows
    Rides = pd.concat([Rides, RidesOneWholeYear])

TS_Data = data.TransformRawDataIntoTSData(Rides)

TS_Data.head()

import hopsworks

Project = hopsworks.login(project = HopsworksProjectName, api_key_value = HopsworksAPIKey)

FeatureStore = Project.get_feature_store()

FeatureGroupName = config.FeatureGroupName
FeatureGroupVersion = config.FeatureGroupVersion

FeatureGroup = FeatureStore.get_or_create_feature_group(name = FeatureGroupName,
                                                        version = FeatureGroupVersion,
                                                        description = "TimeSeries Data at Hourly Frequency",
                                                        primary_key = ["pickup_location_id", "pickup_hour"],
                                                        event_time = "pickup_hour"
                                                       )

FeatureGroup.insert(TS_Data, write_options = {"wait_for_job":False})
